requirements.txt
PyYAML==5.1.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
PyMySQL==0.9.2
SQLAlchemy<1.4
mysql-connector-python==8.0.25
flask-cors==3.0.10